CMS-949-bulk-tagger - splitting classes
