+++ src/Progress/TermManagerExport.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Drupal\dennis_term_manager\Progress;
+
+
+class TermManagerExport {
+
+}
